+++ src/components/internal/index.ts
@@ -2,3 +2,4 @@ export { default as StoryPanel } from './StoryPanel.vue';
 export { default as HelpAndErrorTexts } from './HelpAndErrorTexts.vue';
 export { BlockWrapper, type TBlockWrapperProps } from './BlockWrapper';
 export { ModalBackdropWrapper } from './ModalBackdropWrapper';
+export { default as Loader } from './Loader.vue';