+++ src/components/spartan/SCombobox/SCombobox.vue
@@ -16,6 +16,7 @@ const props = withDefaults(defineProps<Partial<TComboboxProps>>(), {
     modelValue: undefined,
     rounded: 'both',
     displayButtonText: undefined,
+    flipOptions: false,
 });
 
 const store = createContext({ props, emit });
@@ -69,7 +70,7 @@ const store = createContext({ props, emit });
                 <ComboboxOptions
                     v-show="open"
                     static
-                    class="absolute z-10 mt-1 max-h-60 min-w-full overflow-auto rounded-md bg-white text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm"
+                    :class="twMerge('absolute z-10 mt-2 max-h-60 py-1 min-w-full overflow-auto rounded-md bg-white text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm', $props.flipOptions && 'right-0')"
                 >
                     <div
                         v-if="store.emptyResults() && store.query !== ''"