+++ src/components/spartan/SSelect/SSelect.vue
@@ -2,9 +2,10 @@
 import { computed, ref } from 'vue';
 import { roundedClass } from '@/helpers';
 import type { TSelectProps } from './types';
+import { twMerge } from 'tailwind-merge';
 
 const emit = defineEmits<{
-    (e: 'update:modelValue', value: string | undefined): void;
+    (e: 'update:modelValue', value: string | number | undefined): void;
 }>();
 
 const props = withDefaults(defineProps<Partial<TSelectProps>>(), {
@@ -35,11 +36,9 @@ const model = computed({
         v-model="model"
         :disabled="disabled"
         :name="name"
-        :class="[
-            'block border py-2 pl-3 pr-8 text-base text-gray-800',
+        :class="twMerge('block border py-2 pl-3 pr-8 text-base text-gray-800',
             error ? 'border-red-500 focus:s-ring-error' : 'border-gray-300 focus:s-ring',
-            roundedClass[rounded],
-        ]"
+            roundedClass[rounded], props.class)"
     >
         <option v-if="placeholder" disabled :value="undefined">
             {{ placeholder }}