+++ src/components/spartan/STable/STableCell.vue
@@ -1,0 +1,24 @@
+<script setup lang="ts">
+import { twMerge } from 'tailwind-merge';
+
+defineProps<{
+    class?: string;
+    highlight?: boolean;
+    colspan?: number;
+}>();
+</script>
+
+<template>
+    <td
+        :class="
+            twMerge(
+                'p-3.5 text-sm text-gray-500 first:pl-5 last:pr-5',
+                highlight && 'font-medium text-gray-900',
+                $props.class,
+            )
+        "
+        :colspan="colspan"
+    >
+        <slot />
+    </td>
+</template>