+++ src/components/spartan/STable/types.ts
@@ -1,0 +1,14 @@
+
+export type TTableProps = {
+    class: string;
+    borderless: boolean;
+    cols: string[];
+    rows: string[];
+    highlight: number[];
+};
+
+export type TTableHeadCellProps = {
+    class?: string;
+};
+
+export type TTableEmits = {};